Fix grammatical agreement and preserve inflected forms
Enforces strict subject-verb agreement and preserves Greek inflections
throughout the compilation pipeline, maintaining the morphological
integrity of the language.

Changes:
- Add ἔστωσαν (plural imperative) to lexicon for plural subjects
- Preserve inflected forms by using normalize_greek(&original) instead
  of lemma, so ἀριθμοί stays distinct from ἀριθμός
- Update assembler to use normalized original for ordinal/property access
- Cast array indices to usize for Rust compatibility
- Automatically make array bindings mutable for push/pop operations
- Fix collections_demo.glossa to use correct plural agreement

Collections demo now works with proper grammatical agreement catching
errors at "grammar time" - ἀριθμοί ἔστω fails, ἀριθμοί ἔστωσαν succeeds.

+++ src/codegen/rust.rs
@@ -233,7 +233,8 @@ fn generate_expr(expr: &HirExpr) -> TokenStream {
         HirExpr::Index { array, index } => {
             let array_tokens = generate_expr(array);
             let index_tokens = generate_expr(index);
-            quote! { #array_tokens[#index_tokens] }
+            // Cast index to usize for Rust array indexing
+            quote! { #array_tokens[(#index_tokens) as usize] }
         }
 
         HirExpr::Var(name) => {
